requirements.txt
Django==4.2
djangorestframework
gunicorn
pytest
pytest-django
psycopg2-binary
resend
whitenoise